// WADNR.Web/src/app/shared/generated/model/treatment-detail.ts
export class TreatmentDetail { 
    TreatmentID?: number;
    ProjectID?: number;
    ProjectName?: string | null;
    ProjectLocationID?: number | null;
    TreatmentAreaName?: string | null;
    TreatmentTypeID?: number;
    TreatmentTypeName?: string | null;
    TreatmentDetailedActivityTypeID?: number;
    TreatmentDetailedActivityTypeName?: string | null;
    TreatmentCodeID?: number | null;
    TreatmentCodeName?: string | null;
    TreatmentStartDate?: string | null;
    TreatmentEndDate?: string | null;
    TreatmentFootprintAcres?: number;
    TreatmentTreatedAcres?: number | null;
    CostPerAcre?: number | null;
    TotalCost?: number | null;
    TreatmentNotes?: string | null;
    ProgramID?: number | null;
    ProgramName?: string | null;
    ImportedFromGis?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}