// WADNR.Web/src/app/shared/generated/model/treatment-area-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentAreaLookupItem { 
    ProjectLocationID?: number;
    ProjectLocationName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentAreaLookupItemForm { 
    ProjectLocationID?: FormControl<number>;
    ProjectLocationName?: FormControl<string>;
}

export class TreatmentAreaLookupItemFormControls { 
    public static ProjectLocationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectLocationName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}